fix: freehand improvements (still bad)

+++ wasm-src/math/scalar.h
@@ -13,7 +13,7 @@ float clamp(float t, float min, float max);
 float sign(float t);
 float lerp(float a, float b, float t);
 
-bool is_almost_zero(const float t, const float eps);
+bool is_almost_zero(const double t, const float eps);
 bool is_almost_equal(const float t1, const float t2, const float eps);
 bool is_normalized(const float t, bool include_ends);
 bool is_in_range(const float t, const float min, const float max, bool include_ends);
